Set node init script and init check script.
The node init script (node/init.py) sends a custome json operation to the STEEM blockchain to broadcast that the node is live and ready to pin hashes sent to the Nebulus STEEM account.

The init check script (node/networkCheck.py) looks over the past number of operations for any new node init custom json. If found, then adds a line to a file for the any future scripts that need to know the total number of nodes on the network. This may not be needed if we end up not needing the total numebr of nodes.