Refactored pin.py & added pinneedList
pin.py Was refactored to remove unneeded lines of code and to clean up the over all script form clutter. Less imports than were needed befor and the process of pinning and checking pins is now more straight forward.

pinnedList Is where the hashes that get pinned will live until the node sends out the custom json to broadcast the work that was done for payment. Once the script sends the data to the blockchain the file is overwritten.