Updated wedsite landing page copy
index.php now has updated copy to reflect the changes to Nebulus 2.0.0.

node/pin.py now has a commented out line to check the date of the pin for future feature of not pinning content that is x amount of time old. This is to (hopefully) cut the chances of pin scams. The script that will check the chain to pay out nodes will also run a time check and will not pay out on pins older than a certain amount of time.